requirements.txt
Flask
Flask-SQLAlchemy
marshmallow
flask-marshmallow
pytest 
pytest-flask
python-dotenv
